refactoring krl-parser
