+++ packages/krl-parser/src/index.ts
@@ -1,18 +1,5 @@
-import { parse } from "../src/tdop";
+import { parse } from "./krl";
 import tokenizer from "../src/tokenizer";
-const lineColumn = require("line-column");
-const excerptAtLineCol = require("excerpt-at-line-col");
-
-function mkWhere(src: string, line: number, col: number, filename: string) {
-  return {
-    filename,
-    line,
-    col,
-    locationString: filename + ":" + line + ":" + col,
-    excerpt: excerptAtLineCol(src, line - 1, col - 1, 3),
-    excerptOneLine: excerptAtLineCol(src, line - 1, col - 1, 0)
-  };
-}
 
 function parseKRL(src: string, opts?: { filename?: string }) {
   opts = opts || {};
@@ -22,11 +9,8 @@ function parseKRL(src: string, opts?: { filename?: string }) {
     const tree = parse(tokens);
     return tree;
   } catch (e) {
-    if (e && e.token && e.token.loc) {
-      const lc = lineColumn(src, e.token.loc.start);
-      if (lc) {
-        e.where = mkWhere(src, lc.line, lc.col, opts.filename || "");
-      }
+    if (e && e.setupWhere) {
+      e.setupWhere(src, opts.filename || "");
     }
     throw e;
   }
